requirements.txt
pyjwt==2.6.0
Django==4.1.7
gunicorn